Initial commit of SVG Widgets for iPython notebook
This is a library of Scalar Vector Graphic (SVG) Widgets for the
iPython notebook (versions 2.0 onwards). Widgets allow the
synchronisation of Widgets between the Python backend and Javascript
frontend. This library implements this functionality for SVG elements.

This basic version only works for groups <g> and rectangles <rect>.

Changes to be committed:
	new file:   svgwidgets.js
	new file:   svgwidgets.py
	new file:   svgwidgets_example.ipynb